finish

+++ src/main/java/com/pruebatecnica/picpay/domain/dto/request/TransferRequestDto.java
@@ -1,7 +1,9 @@
 package com.pruebatecnica.picpay.domain.dto.request;
 
+import java.math.BigDecimal;
+
 public record TransferRequestDto(
-        double amount,
+        BigDecimal amount,
         Long cardSender,
         Long cardReceiver
 ) {
